requirements.txt
autoload-module==2.0.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
idna==3.10
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-utils==3.9.0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.3.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
